requirements.txt
google-cloud-aiplatform
google-cloud-documentai
langchain
langchain-community
streamlit